move data_reader to data/ package; cleanup logging from json file reader
