+++ data/data_reader.go
@@ -1,9 +1,8 @@
-package main
+package data
 
 import (
 	"encoding/json"
 	"io"
-	"log"
 	"os"
 )
 
@@ -15,20 +14,17 @@ type Sample struct {
 func ReadJson(jsonFilePath string) (int64, []string, error) {
 	jsonFile, err := os.Open(jsonFilePath)
 	if err != nil {
-		log.Fatalf("Failed to open file: %v", err)
 		return 0, nil, err
 	}
 	defer jsonFile.Close()
 
 	byteValue, err := io.ReadAll(jsonFile)
 	if err != nil {
-		log.Fatalf("Failed to read file: %v", err)
 		return 0, nil, err
 	}
 
-	var dataContent DataContent
+	var dataContent Sample
 	if err := json.Unmarshal(byteValue, &dataContent); err != nil {
-		log.Fatalf("Failed to unmarshal JSON: %v", err)
 		return 0, nil, err
 	}
